Extract relative time component for reuse
