+++ frontend/src/app/features/in-app-notifications/entry/actors-line/in-app-notification-actors-line.component.sass
@@ -7,11 +7,6 @@
   align-items: center
   color: var(--fgColor-muted)
 
-  &--date
-    @include text-shortener
-    max-width: 100%
-    line-height: 1rem
-
   &--container
     @include text-shortener
     display: flex
