+++ frontend/src/app/features/in-app-notifications/entry/relative-time/in-app-notification-relative-time.component.html
@@ -1,0 +1,5 @@
+<div
+  class="op-ian-relative-time"
+  [title]="fixedTime"
+  [textContent]="relativeTime$ | async">
+</div>
