+++ frontend/src/app/features/in-app-notifications/entry/relative-time/in-app-notification-relative-time.component.sass
@@ -1,0 +1,6 @@
+@import "helpers"
+
+.op-ian-relative-time
+  @include text-shortener
+  max-width: 100%
+  line-height: 1rem
